Actualizacion endpoints
